# Conflicts:
#	core/src/main/java/com/StardewValley/Controllers/Camera.java
#	core/src/main/java/com/StardewValley/Controllers/GameController.java
#	core/src/main/java/com/StardewValley/Controllers/PlayerController.java
#	core/src/main/java/com/StardewValley/Controllers/RegistrationController.java
#	core/src/main/java/com/StardewValley/Controllers/WordController.java
#	core/src/main/java/com/StardewValley/Main.java
#	core/src/main/java/com/StardewValley/Views/GameMenu.java
#	core/src/main/java/com/StardewValley/Views/GameView.java
#	core/src/main/java/com/StardewValley/Views/LoginMenu.java
#	core/src/main/java/com/StardewValley/Views/PickQuestion.java
#	core/src/main/java/com/StardewValley/Views/RegistrationMenu.java

+++ core/src/main/java/com/StardewValley/Models/Player.java
@@ -136,7 +136,7 @@ public class Player implements Serializable {
             this.energy.addEnergy(amount);
     }
 
-    public void decreaseEnergy(int amount) {
+    public void decreaseEnergy(float amount) {
         this.energy.decreaseEnergy(amount);
     }
